Fix an overflow issue in 4 byte to unsigned integer conversion (needed to be a long). Fix a related TODO in ConnectionSettings
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1682984 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/coyote/http2/ByteUtil.java
@@ -68,8 +68,8 @@ public class ByteUtil {
     }
 
 
-    public static int getFourBytes(byte[] input, int firstByte) {
-        return ((input[firstByte] & 0xFF) << 24) + ((input[firstByte + 1] & 0xFF) << 16) +
+    public static long getFourBytes(byte[] input, int firstByte) {
+        return ((long)(input[firstByte] & 0xFF) << 24) + ((input[firstByte + 1] & 0xFF) << 16) +
                 ((input[firstByte + 2] & 0xFF) << 8) + (input[firstByte + 3] & 0xFF);
     }
 }
